requirements.txt
Django==2.0
django-admin-rangefilter==0.3.6
django-admin-relation-links==0.0.5
mysqlclient==1.3.12
pytz==2018.4
